sync

+++ library_dicom/dicom_processor/tools/threshold_mask.py
@@ -10,38 +10,38 @@ def threshold_matrix(mask_array, pet_array, threshold):
             number_of_roi = mask_array.shape[3]
             liste = []
             for i in range(number_of_roi):
-                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] ))
+                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] )).astype('uint8')
                 suv_values = []
                 x,y,z = np.where(mask_array[:,:,:,i] != 0)
                 if len(x) != 0 : 
                     suv_values = pet_array[x,y,z].tolist()
                     seuil = np.max(suv_values) * threshold
                     new_mask[np.where((pet_array > seuil) & (mask_array[:,:,:,i] > 0))] = 1
-                    liste.append(new_mask)
-                else : liste.append(new_mask)
-            return np.stack(liste, axis = 3)    
+                    liste.append(new_mask.astype('uint8'))
+                else : liste.append(new_mask.astype('uint8'))
+            return np.stack(liste, axis = 3).astype('uint8')   
 
         else : 
             maxi = np.max(mask_array)
             if maxi == 1.0 : 
-                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] ))
+                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] )).astype('uint8')
                 x,y,z = np.where(mask_array != 0)
                 suv_values = pet_array[x,y,z].tolist()
                 seuil = np.max(suv_values) * threshold
                 new_mask[np.where((pet_array > seuil) & (mask_array > 0))] = 1
-                return new_mask
+                return new_mask.astype('uint8')
                  
 
             else :  
                 number_of_roi = maxi
-                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] ))
+                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] )).astype('uint8')
                 for i in range(number_of_roi):
                     suv_values = []
                     x,y,z = np.where(mask_array == i)
                     suv_values = pet_array[x,y,z].tolist()
                     seuil = np.max(suv_values) * threshold
                     new_mask[np.where((pet_array > seuil) & (mask_array == i))] = i
-                return new_mask
+                return new_mask.astype('uint8')
 
 
 
@@ -50,7 +50,7 @@ def threshold_matrix(mask_array, pet_array, threshold):
             number_of_roi = mask_array.shape[3]
             liste = []
             for i in range(number_of_roi):
-                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] ))
+                new_mask = np.zeros((mask_array.shape[0],mask_array.shape[1], mask_array.shape[2] )).astype('uint8')
                 seuil = threshold
                 new_mask[np.where((pet_array > seuil) & (mask_array[:,:,:,i] > 0))] = 1
                 liste.append(new_mask)
